// client/src/app/modules/quotation/model/BodaccSale.ts
import { CompetentAuthority } from "../../miscellaneous/model/CompetentAuthority";
import { LegalForm } from "../../miscellaneous/model/LegalForm";
import { ActType } from "./ActType";
import { FundType } from "./FundType";

export interface BodaccSale {
  id: number;
  divestedBusinessAddress: string;
  fundType: FundType;
  divestedBusinessActivities: string;
  isOwnerIndividual: boolean;
  ownerFirstname: string;
  ownerLastname: string;
  ownerDenomination: string;
  ownerSiren: string;
  ownerAddress: string;
  ownerShareCapital: number;
  ownerAbbreviation: string;
  ownerBusinessName: string;
  ownerLegalForm: LegalForm;
  purchaserFirstname: string;
  purchaserLastname: string;
  purchaserSiren: string;
  purchaserDenomination: string;
  purchaserLegalForm: LegalForm;
  purchaserAbbreviation: string;
  purchaserBusinessName: string;
  purchaserShareCapital: number;
  purchaserMinimumCapital: number;
  purchaserActivityStartDate: Date;
  leaseResilisationDate: Date;
  leaseResilisationFee: number;
  leaseAddress: string;
  deedDate: Date;
  salePrice: number;
  registrationDate: Date;
  registrationAuthority: CompetentAuthority;
  registrationReferences: string;
  actType: ActType;
  writor: string;
  writorAddress: string;
  validityObjectionAddress: string;
  mailObjectionAddress: string;
  isTenantIndividual: boolean;
  tenantFirstname: string;
  tenantLastname: string;
  tenantDenomination: string;
  tenantSiren: string;
  tenantAddress: string;
  tenantShareCapital: number;
  tenantMinimumCapital: number;
  tenantAbbreviation: string;
  tenantBusinessName: string;
  tenantLegalForm: LegalForm;
}